refactor: move commands to a separate commands.go file
