+++ cmd/client/main.go
@@ -37,37 +37,6 @@ type model struct {
 	err         error
 }
 
-func listenFromConn(conn *websocket.Conn, sub chan<- chat.Message) tea.Cmd {
-	return func() tea.Msg {
-		var data chat.Message
-		for {
-
-			if err := conn.ReadJSON(&data); err != nil {
-				return errorMsg(err)
-			}
-			sub <- data
-		}
-	}
-}
-
-func waitForMessage(sub <-chan chat.Message) tea.Cmd {
-	return func() tea.Msg {
-		message := <-sub
-		return receivedMsg(message)
-	}
-}
-
-func writeToConn(conn *websocket.Conn, message string) tea.Cmd {
-	return func() tea.Msg {
-		err := conn.WriteMessage(websocket.TextMessage, []byte(message))
-		if err != nil {
-			return errorMsg(err)
-		}
-
-		return nil
-	}
-}
-
 func initialModel(conn *websocket.Conn) model {
 	ta := textarea.New()
